refactor : exception type -> exception message, 메일 인증 코드 불일치 메세지 추가
- Type 명칭보다는 Message 명칭이 좀 더 어울리는 것 같아서 수정
- 메일 인증 코드 불일치 시 출력 문구 변경 : "인증 코드 만료" -> "인증 코드 불일치"

issue key : SB2-7
